Refactoring Data provider model classes
- Renaming the packages.
- Addressing few review comments.

Part of #468
