+++ src/main/java/com/sap/oss/phosphor/fosstars/data/artifact/experimental/owasp/data/Software.java
@@ -1,0 +1,29 @@
+package com.sap.oss.phosphor.fosstars.data.artifact.experimental.owasp.data;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import java.io.Serializable;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonPropertyOrder({"versionStartIncluding", "versionEndIncluding"})
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Software implements Serializable {
+
+  @JsonProperty("versionStartIncluding")
+  private String versionStartIncluding;
+
+  @JsonProperty("versionEndIncluding")
+  private String versionEndIncluding;
+
+  @JsonProperty("versionStartIncluding")
+  public String getVersionStartIncluding() {
+    return versionStartIncluding;
+  }
+
+  @JsonProperty("versionEndIncluding")
+  public String getVersionEndIncluding() {
+    return versionEndIncluding;
+  }
+}
