+++ src/main/java/com/sap/oss/phosphor/fosstars/data/artifact/experimental/owasp/data/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * The package contains data providers which fetch data from OWASP Dependency-Check.
+ */
+package com.sap.oss.phosphor.fosstars.data.artifact.experimental.owasp.data;
